+++ Changelog.md
@@ -1,6 +1,7 @@
 # 11.0.2-20240921-2
 * Don't hardcode Accept and Cancel.
 * Persist window size across logins.
+* Add config button to reset window size and position.
 
 # 11.0.2-20240921-1
 * Fix minimap icon disable/enable.